requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.11
sqlalchemy==1.4.49
alembic==1.11.1
pytesseract==0.3.10
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.0.0
itsdangerous==2.1.2
requests==2.31.0
jinja2==3.1.2
python-jose==3.3.0
psycopg2-binary==2.9.9
